// src/components/solicitudes-pago/solicitudes-pago.types.tsx
export type CrearSolicitudNominaIndividualPayload = {
  tipo_solicitud: "PAGO_NOMINA";
  modalidad_nomina: "INDIVIDUAL";
  periodo_nomina: string;
  proyecto_base_id: string;
  centro_costo_id: string;
  beneficiario_id: string;
  concepto_nomina: string;
  medio_pago: MedioPagoSolicitud;
  descripcion: string;
  valor_bruto: number;
  valor_retenciones: number;
  valor_descuentos: number;
};

export type CrearSolicitudFrontendPayload =
  | CrearSolicitudProveedorPayload
  | CrearSolicitudNominaIndividualPayload;

export type NominaIndividualFormularioState = {
  proyecto_base_id: string;
  centro_costo_id: string;
  beneficiario_id: string;
  periodo_nomina: string;
  concepto_nomina: string;
  medio_pago: MedioPagoSolicitud | "";
  descripcion: string;
  valor_bruto: string;
  valor_retenciones: string;
  valor_descuentos: string;
};